Rename gated content vars and files
