+++ packages/discovery-provider/src/gated_content/helpers.py
@@ -3,7 +3,7 @@ from typing import Dict, Union
 
 from sqlalchemy.orm.session import Session
 
-from src.gated_content.gated_content_types import GatedContentType
+from src.gated_content.types import GatedContentType
 from src.models.social.follow import Follow
 from src.models.users.aggregate_user_tips import AggregateUserTip
 from src.models.users.usdc_purchase import USDCPurchase
