+++ packages/discovery-provider/src/gated_content/types.py
