+++ packages/discovery-provider/src/queries/get_trending.py
@@ -1,9 +1,7 @@
 import logging
 
 from src.api.v1.helpers import extend_track, format_limit, format_offset
-from src.gated_content.gated_content_constants import (
-    SHOULD_TRENDING_EXCLUDE_PREMIUM_TRACKS,
-)
+from src.gated_content.constants import SHOULD_TRENDING_EXCLUDE_PREMIUM_TRACKS
 from src.queries.generate_unpopulated_trending_tracks import TRENDING_TRACKS_LIMIT
 from src.queries.get_trending_tracks import get_trending_tracks
 from src.utils.helpers import decode_string_id  # pylint: disable=C0302
